// src/models/TransactionData.ts
export type TransactionData = {
  id: string;
  transaction_code: string;
  transaction_status: string;
  transaction_date: string;
  producer: string;
  product_id: string;
  product_name: string;
  offer_id: string;
  offer_name: string;
  currency: string;
  purchase_value_with_tax: number;
  purchase_value_without_tax: number;
  commission_currency: string;
  my_commission_value: number;
  src_code: string;
  sck_code: string;
  payment_method: string;
  total_installments: number;
  total_charges: number;
  coupon_code: string;
  buyer_name: string;
  buyer_email: string;
  buyer_country: string;
  buyer_phone: string;
  buyer_document: string;
  buyer_state: string;
  buyer_instagram: string;
  order_bump_type: "Parent" | "Child" | "(none)";
  order_bump_index: "Child" | null;
  order_bump_transaction: string;
  user_id: string;
  project_id: string;
  platform: string;
};